impl-submit-report - move operation tests into tests/bark/operations
