+++ tests/bark/operations/op-fetch-report.test.ts
@@ -1,17 +1,14 @@
 import { BarkReport, BarkReportData } from "@/lib/bark/bark-models";
-import { givenDog, givenVet } from "./_given";
-import { withBarkContext } from "./_context";
-import { mockReportData } from "./_mocks";
-import { CODE } from "@/lib/utilities/bark-code";
+import { givenDog, givenVet } from "../_given";
+import { mockReportData } from "../_mocks";
+import { withBarkContext } from "../_context";
 import { opRecordAppointmentCallOutcome } from "@/lib/bark/operations/op-record-appointment-call-outcome";
 import { opSubmitReport } from "@/lib/bark/operations/op-submit-report";
 import { opFetchReport } from "@/lib/bark/operations/op-fetch-report";
-import { opEditReport } from "@/lib/bark/operations/op-edit-report";
 
-describe("opEditReport", () => {
-  it("should update report data to the new values", async () => {
+describe("opFetchReport", () => {
+  it("should return the report", async () => {
     await withBarkContext(async ({ context, testContext }) => {
-      // GIVEN an existing report
       const { vetId } = await givenVet(testContext);
       const { dogId } = await givenDog(testContext, { preferredVetId: vetId });
       const res1 = await opRecordAppointmentCallOutcome(context, {
@@ -19,31 +16,16 @@ describe("opEditReport", () => {
         vetId,
       });
       const { appointmentId } = res1.result!;
-      const originalReportData: BarkReportData = {
+      const reportData: BarkReportData = {
         ...mockReportData(),
         ineligibilityReason: "test reason",
-        dogBodyConditioningScore: 1,
       };
       const res2 = await opSubmitReport(context, {
         appointmentId,
-        reportData: originalReportData,
+        reportData,
       });
       const { reportId } = res2.result!;
       console.log({ reportId });
-
-      // WHEN the report is modified
-      const modifiedReportData: BarkReportData = {
-        ...originalReportData,
-        ineligibilityReason: "new reason",
-        dogBodyConditioningScore: 10,
-      };
-      const res3 = await opEditReport(context, {
-        reportId,
-        reportData: modifiedReportData,
-      });
-      expect(res3).toEqual(CODE.OK);
-
-      // THEN expect the modifications to be persisted.
       const { result, error } = await opFetchReport(context, {
         reportId,
       });
@@ -56,7 +38,7 @@ describe("opEditReport", () => {
         reportId,
         reportCreationTime: report.reportCreationTime,
         reportModificationTime: report.reportModificationTime,
-        ...modifiedReportData,
+        ...reportData,
       };
       expect(report).toEqual(expectedReport);
     });
